requirements.txt
selenium
scrapy
daiquiri
collections
transformers
pytorch
numpy
pandas
dataclasses
typing
